+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,7 @@ upcoming
 ++++++
 * 'az containerapp debug': Open an SSH-like interactive shell within a container app debug console.
 * 'az containerapp env java-component': Support more flexible configuration updates with new parameters --set-configurations, --replace-configurations, --remove-configurations and --remove-all-configurations.
+* 'az containerapp env java-component gateway-for-spring': Support --bind --unbind in create/update.
 
 1.1.0b1
 ++++++